requirements.txt
arxiv>=3.0.0
PyYAML>=6.0.1
pdfplumber>=0.11.0
requests>=2.31.0
